+++ src/com/geekbrains/javaCoreAdvancedLevel/lesson_01/task1/Robot.java
@@ -1,49 +1,8 @@
 package com.geekbrains.javaCoreAdvancedLevel.lesson_01.task1;
 
-import com.geekbrains.javaCoreAdvancedLevel.lesson_01.task2.HavingSportParam;
-import com.geekbrains.javaCoreAdvancedLevel.lesson_01.task2.RunningTrack;
-import com.geekbrains.javaCoreAdvancedLevel.lesson_01.task2.Wall;
-
-public class Robot implements SuperAction{
-    private String name;
-    private int maxJump;
-    private int maxDist;
-
-    public Robot(String name) {
-        this.name = name;
-    }
+public class Robot extends MaterialObject{
 
     public Robot(String name, int maxJump, int maxDist) {
-        this.name = name;
-        this.maxJump = maxJump;
-        this.maxDist = maxDist;
-    }
-
-    @Override
-    public boolean jump(HavingSportParam wall) {
-        if (wall instanceof Wall){
-            System.out.println("Robot " + this.name + (wall.getParam() > this.maxJump? " did not jumped over wall " : " jumped over wall ") + wall.getParam() + " meters high");
-            return wall.getParam() > this.maxJump;
-        }
-        return false;
-    }
-
-    @Override
-    public void jump() {
-        System.out.println("Robot " + this.name + " jump");
-    }
-
-    @Override
-    public boolean run(HavingSportParam track) {
-        if (track instanceof RunningTrack){
-            System.out.println("Robot " + this.name + (track.getParam() > this.maxDist ? " did not ran a distance of " : " ran a distance of ") + track.getParam() + " meters");
-            return track.getParam() > this.maxDist;
-        }
-        return false;
-    }
-
-    @Override
-    public void run() {
-        System.out.println("Robot " + this.name + " run");
+        super(name, maxJump,maxDist);
     }
 }
