+++ src/com/geekbrains/javaCoreAdvancedLevel/lesson_01/task1/Running.java
@@ -1,6 +1,5 @@
 package com.geekbrains.javaCoreAdvancedLevel.lesson_01.task1;
 
-
 public interface Running {
     void run();
 }
