+++ src/com/geekbrains/javaCoreAdvancedLevel/lesson_01/task1/SuperAction.java
@@ -1,6 +1,6 @@
 package com.geekbrains.javaCoreAdvancedLevel.lesson_01.task1;
 
-import com.geekbrains.javaCoreAdvancedLevel.lesson_01.task2.HavingSportParam;
+import com.geekbrains.javaCoreAdvancedLevel.lesson_01.task2_3_4.HavingSportParam;
 
 public interface SuperAction extends Running, Leaping{
     boolean run(HavingSportParam track);
