--- src/com/geekbrains/javaCoreAdvancedLevel/lesson_01/task2/HavingSportParam.java
@@ -1,5 +1,0 @@
-package com.geekbrains.javaCoreAdvancedLevel.lesson_01.task2;
-
-public interface HavingSportParam {
-    int getParam();
-}
