+++ src/com/geekbrains/javaCoreAdvancedLevel/lesson_01/task2_3_4/HavingSportParam.java
@@ -1,0 +1,5 @@
+package com.geekbrains.javaCoreAdvancedLevel.lesson_01.task2_3_4;
+
+public interface HavingSportParam {
+    int getParam();
+}
